Rename resources package to rest
