+++ server/src/main/java/com/krillsson/sysapi/rest/ProcessesResource.java
@@ -18,7 +18,7 @@
  * Maintainers:
  * contact[at]christian-jensen[dot]se
  */
-package com.krillsson.sysapi.resources;
+package com.krillsson.sysapi.rest;
 
 
 import com.krillsson.sysapi.auth.BasicAuthorizer;
