+++ server/src/test/java/com/krillsson/sysapi/rest/CpuResourceTest.java
@@ -1,4 +1,4 @@
-package com.krillsson.sysapi.resources;
+package com.krillsson.sysapi.rest;
 
 import com.krillsson.sysapi.core.domain.cpu.CpuInfo;
 import com.krillsson.sysapi.core.history.MetricsHistoryManager;
